+++ app/services/members/members_service.rb
@@ -36,10 +36,12 @@ class Members::MembersService
       end
     end
 
-    if params[:invoicing_profile_attributes][:organization] == 'false'
-      params[:invoicing_profile_attributes].reject! { |p| %w[organization_attributes organization].include?(p) }
-    else
-      params[:invoicing_profile_attributes].reject! { |p| p == 'organization' }
+    if params[:invoicing_profile_attributes] && params[:invoicing_profile_attributes][:organization]
+      if params[:invoicing_profile_attributes][:organization] == 'false'
+        params[:invoicing_profile_attributes].reject! { |p| %w[organization_attributes organization].include?(p) }
+      else
+        params[:invoicing_profile_attributes].reject! { |p| p == 'organization' }
+      end
     end
 
     not_complete = member.need_completion?