+++ test/integration/members/as_admin_test.rb
@@ -100,4 +100,23 @@ class MembersTest < ActionDispatch::IntegrationTest
 
     assert_match(/Kevin/, res[0][:name])
   end
+
+  test 'admin changes the group of a member' do
+    user = User.find(2)
+    patch "/api/members/#{user.id}/",
+          params: {
+            user: {
+              id: user.id,
+              group_id: 2
+            }
+          }
+
+    # Check response format & status
+    assert_equal 200, response.status, response.body
+    assert_equal Mime[:json], response.content_type
+
+    # Check search result
+    res = json_response(response.body)
+    assert_equal 2, res[:group_id]
+  end
 end