+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Display bookers names to connected users now apply to all resources
 - Fix a bug: unable to book a space's slot with an existing reservation
 - Fix a bug: Unable to import accounts from SSO when the transformation modal was opened but leaved empty
+- Fix a bug: Unable to change the group of a user
 
 ## v5.4.12 2022 July 06